# crates/recrypt-client/openapi.json
{
  "components": {
    "schemas": {
      "AccountResponse": {
        "description": "Response body for `POST /accounts` and `GET /accounts/{fingerprint}`.",
        "properties": {
          "created_at": {
            "description": "Account creation time (Unix seconds).",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "ed25519_pk": {
            "description": "Base58-encoded 32-byte Ed25519 public key (echoed from the request).",
            "type": "string"
          },
          "fingerprint": {
            "description": "Base58-encoded BLAKE3 fingerprint of the Ed25519 public key. Used\nas the account's stable identifier in URL path segments.",
            "type": "string"
          },
          "ml_dsa_pk": {
            "description": "Base64-encoded ML-DSA-87 public key (echoed from the request).",
            "type": "string"
          }
        },
        "required": [
          "fingerprint",
          "ed25519_pk",
          "ml_dsa_pk",
          "created_at"
        ],
        "type": "object"
      },
      "CreateAccountRequest": {
        "description": "Request body for `POST /accounts`.",
        "properties": {
          "ed25519_pk": {
            "description": "Base58-encoded 32-byte Ed25519 public key. Base58 is retained\nhere for parity with `fingerprint`, which is derived from this\nkey and used in URL path segments. Tracked under recrypt-nj1\nfollow-up: unify body-byte encoding to base64.",
            "example": "C8X8K9...32B base58...",
            "type": "string"
          },
          "ml_dsa_pk": {
            "description": "Base64-encoded ML-DSA-87 public key (~2.6 KB).",
            "example": "MIIChw...base64...",
            "type": "string"
          }
        },
        "required": [
          "ed25519_pk",
          "ml_dsa_pk"
        ],
        "type": "object"
      },
      "VerifyCapabilityRequest": {
        "description": "Request body for `POST /capabilities/verify`.",
        "properties": {
          "envelope_b64": {
            "description": "Base64-encoded Gordian Envelope (dCBOR) of a signed capability.\nThe wire form produced by `Capability::sign`.",
            "type": "string"
          },
          "issuer_ed25519_pk": {
            "description": "Base64-encoded 32-byte Ed25519 public key the capability was\nsigned with. Lookup hint: derive from the issuer fingerprint\nvia `GET /accounts/{fingerprint}`.",
            "type": "string"
          },
          "issuer_ml_dsa_pk": {
            "description": "Base64-encoded ML-DSA-87 public key (~2.6 KB), if the\ncaller wants to require post-quantum signature verification.\nOmitted = classical-only check.",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "envelope_b64",
          "issuer_ed25519_pk"
        ],
        "type": "object"
      },
      "VerifyCapabilityResponse": {
        "description": "Response body for `POST /capabilities/verify`.\n\n`valid: true` means the signature checked out and the capability\nhas not expired. Permission set, expiry, and parent-chain link are\nreturned for the caller to enforce against their own policy.",
        "properties": {
          "expired": {
            "type": "boolean"
          },
          "expires_at": {
            "format": "int64",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "granted_to": {
            "description": "Base58 fingerprint of the recipient.",
            "type": "string"
          },
          "issuer": {
            "description": "Base58 fingerprint of the issuer.",
            "type": "string"
          },
          "note": {
            "nullable": true,
            "type": "string"
          },
          "parent": {
            "description": "Base58 of the parent capability's wrapped envelope digest, when\npresent. Chain verification is not performed by this endpoint\n(see recrypt-91h follow-up).",
            "nullable": true,
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "subject": {
            "description": "Base58 of the 32-byte resource address.",
            "type": "string"
          },
          "subject_kind": {
            "description": "`\"file\"`, `\"keyspace\"`, or `\"account\"`.",
            "type": "string"
          },
          "valid": {
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "expired",
          "subject_kind",
          "subject",
          "granted_to",
          "issuer",
          "permissions"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Recrypt proxy HTTP API. Endpoints are migrated to schema-as-source-of-truth incrementally (epic recrypt-nj1).",
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "title": "recrypt",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/accounts": {
      "post": {
        "description": "Registers the caller's dual-stack public keys (Ed25519 + ML-DSA-87)\nunder the BLAKE3 fingerprint of their Ed25519 key. Subsequent\nauthenticated endpoints look up this account record to verify\nper-request multisigs.\n\n**Authorization**: dual-stack multisig over the canonical message\n`CREATE:{ed25519_pk}:{ml_dsa_pk}:{nonce}` using the very keys being\nregistered. The `X-Public-Key` header MUST equal the fingerprint\nderived from `ed25519_pk`; otherwise the request is rejected as\nmalformed.",
        "operationId": "create_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "Account created"
          },
          "400": {
            "description": "Malformed request or fingerprint mismatch"
          },
          "409": {
            "description": "Account already exists"
          }
        },
        "summary": "Create a new account.",
        "tags": [
          "accounts"
        ]
      }
    },
    "/capabilities/verify": {
      "post": {
        "operationId": "verify_capability",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCapabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyCapabilityResponse"
                }
              }
            },
            "description": "Capability parsed and signature verified (check `expired`)"
          },
          "400": {
            "description": "Malformed envelope or key encodings"
          },
          "401": {
            "description": "Signature verification failed"
          }
        },
        "summary": "Verify a presented capability token.",
        "tags": [
          "capabilities"
        ]
      }
    }
  },
  "tags": [
    {
      "description": "Account registration and lookup.",
      "name": "accounts"
    },
    {
      "description": "Bearer-token (Capability) verification.",
      "name": "capabilities"
    }
  ]
}